Makefile: factorize out the gcc rules for gcc >= 3
As make supports variable names determined via evaluation, it becomes
possible to create associative arrays so that we keep a gcc version
number for each gcc major version. This simplifies the configuration
a lot since there is no rule duplication, version tracking or if/else
anymore.

Gcc 2.95 is still independant because it is very different. The gcc
used to build the glibc is now designated by the variable GCCLIBC
which must point to a gcc branch version (eg: 2.95).