patches: add a ct-ng patch to fix usage of "eabi" name in ARM
This patch is temporarily disabled. Its principle is to remove the
"eabi" suffix from the target name for the time it takes to adapt the
build scripts which do not recognize this as a valid arch since it
does not match "<arch>-<os>". It should disappear when tools are
fixed for this.