toolchain: major update and reorganization
This is a major update.

The config files were renamed, there were too many conficts, all
experimental ARM files became unmaintainable. All old or experimental
config files were moved to the "oldconfig" subdirectory.

The "patches/" directory was reorganized. Now pure ct-ng patches are stored
under "ct-ng-build/$VER", while other components patches to add to ct-ng are
stored under "ct-ng-patches/$COMPONENT/$VER/". That way the script is able
to automatically scan the appropriate directories and apply all patches found.
Patches names with a leading dot are ignored.

The list of ct-ng patches to remove is not in the script anymore but listed
in the "patches/to-remove" file.

Instead of simply passing versions in the environment, the script now
retrieves whatever it can from the config files. This allows it to safely
iterate over a set of config files and even to rely on different ct-ng
versions for multiple toolchains.

As a result, the script now iterates over a number of actions specified on
the command line instead of hard-set in the code. There's no configuration
in the code anymore, everything is found from environment, configs and
directories.

For now, building extra tools doesn't work anymore.