toolchain: allow multiple configs to use different ct-ng versions
Glibc 2.3 only builds with 1.7.2. Last possible version is 1.9.3 but
it breaks other things. Newer versions are the only ones allowing
more recent compilers / libcs. So we need to support a wide variety
of ct-ng versions.

Each of them is now built in its own directory, and the toolchains
are now built in their own dir too (not build/tc anymore), so that
parallel builds are much more possible.

GDB and should still be removed from the configs or statically built,
because it doesn't work since it was linked with the dynamic libgcc_s.so.
Ideally the build scripts should be changed so that libgcc_s is removed
just after packaging.