patches/ct-ng: add a CT_ARCH_OVERRIDE variable to 1.15.3
This variable allows to force the arch component in the resulting
toolchain name, which is sometimes needed to differenciate between
multiple variants of the same architecture (eg: armv5 vs armv7).

Note that the new name must still be recognized as valid by all
the tools, which generally is true if it starts with the radix
name.