toolchain: fix __sched_setaffinity() in glibc 2.9
This one would randomly crash (in fact it would crash whenever the pid
is not multiple of 4). The reason is that the way it loads getpid() in
a macro does not ensure argument ordering, and with gcc 4.7, it's loaded
after register r1 is set, causing it to be clobbered by getpid() since
ARM's EABI does not protect r0..r3.

So the INTERNAL_SYSCALL macro must not be called with a function call
in any of its arguments. The fix consists in using a temporary variable
on the stack.