config: add config for PJ4B in thumb mode with glibc-2.17+uclibc-0.9.33.2
Glibc 2.9 doesn't build in thumb mode, too many fixes are need in
the asm code. 2.17 builds out of the box. A few fixes were needed
for uclibc.

This config also switches to gdb 7.4.1, since 6.8a does not build
with recent glibc (issues with siginfo).

This config is set by default for a pj4b CPU, which enables the div
instruction in thumb mode only.