config: add config for PJ4B in ARM mode with glibc-2.17+uclibc-0.9.33.2
Since glibc was switched to 2.17 for thumb mode, switch it to 2.17
for ARM mode as well.

This config also switches to gdb 7.4.1, since 6.8a does not build
with recent glibc (issues with siginfo).

This config is set by default for a pj4b CPU, so it does not have
the div instruction enabled by default.