config: arm-2.18: min kernel version is 2.6.32, not 2.6.32.62 nor .60
By default the toolchain assumes that we'll be using the same version
as found in the kernel headers. But doing so, we risk not being able
to run executables on older maintenance kernels. Instead we want to
run on the base version of a maintenance kernel. The proper way to
achieve this is by setting CT_LIBC_GLIBC_KERNEL_VERSION_CHOSEN=y
and setting the minimum kernel version in CT_LIBC_GLIBC_MIN_KERNEL
and CT_LIBC_GLIBC_MIN_KERNEL_VERSION. Note that ct-ng automatically
trims the 4th digit of the version from CT_LIBC_GLIBC_MIN_KERNEL_VERSION.

Note that ct-ng also accepts versions passed as --enable-kernel=<version>
in CT_LIBC_GLIBC_EXTRA_CONFIG_ARRAY.