toolchain: use native 'ar' and 'ranlib' instead of target's
During a canadian cross build, toolchain experiences a failure when
trying to extract libgcc_s.a because it uses the freshly built 'ar'
and 'ranlib' utilities which are not necessarily compatible with the
build host's. Let's use the native ones instead and allow to override
them using environment variables just like for make.