toolchain: keep a copy of the list of glibc patches during the build
Currently it is problematic to know if the glibc shipped with products
is up to date or not. Now we emit the list of all patches applied to
glibc into the output toolchain directory, so that a script responsible
for building a glibc package can verify that no patch is missing.