toolchain: also allow to force CT_BUILD
Some machines support multiple native archs (eg: i586 / x86_64,
aarch64 / armthf). It is sometimes necessary to be able to force
the architecture to use to build some tools during the process.
The new TC_BUILD does this and works similarly to TC_HOST except
that it focuses on earlier steps. It may still fail with the
companion tools. At least automake was noticed not using the
proper local compiler and trying to apply CT_CC_CFLAGS_FOR_BUILD
to it.