config: enable CT_GDB_NATIVE_STATIC in all recent configs except x86_64
Some configs used to have it, some not. Let's make all this more
homogenous. However, x86_64 doesn't support it because on this
platform gcc complains when called with -static and -shared at the same
time. While this is a perfectly well documented combination, gcc still
refuses to support it on x86_64. It also looks odd to see gdb try to
produce libinproctrace.so when asked to produce a static binary.