toolchain: automatically build libatomic for architectures that require it
In gcc-4.7, some atomic ops cannot always be implemented natively, and a
call is made to the equivalent function. The problem is, libatomic only
appeared in 4.8. The gcc folks provide a simple implementation on the gcc
site (https://gcc.gnu.org/wiki/Atomic/GCCMM) that can be linked with any
other version, assuming a little bit of defines and accepting a few warnings.

What this patch does is to check if libatomic is needed by checking if the
final stage compiler may emit such calls (any *_LOCK_FREE macro is defined
to 1). If so it checks if a libatomic is present in the sysroot or the
compiler's directory, otherwise it will use the final stage compiler for
the build architecture to compile it, and will store it into the compiler's
directory next to libgcc_s.a. A few tricks are used for canadian builds such
as being extra careful to use the correct compiler!

It's mainly MIPS on gcc-4.7 that benefits from this, but the compiler remains
useful for plenty of projects.