config: add a riscv64 config for gcc-11.4, binutils-2.38 and glibc-2.31
This one relies on crosstool-ng-1.26.0.

It managed to build a bootable riscv64 kernel 5.10.215 for licheerv-nano
as well as the preinit based on nolibc.

It's worth noting that there are different vectors implementations for
this platform and that the most common ones are not present in upstream
binutils (v0.7), but the more recent ones are (v1.0). Also some
instructions such as vsetvli were missing from binutils-2.37's gas, so
it looks like 2.38 is the first properly working version for this
platform.